requirements.txt
openai
python-dotenv
Pillow
pathlib